requirements.txt
#requirements.txt

python-dotenv
requests
pandas
matplotlib